% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hla_class1.R
\name{clean_hla_class1}
\alias{clean_hla_class1}
\title{Clean and standardize messy HLA Class I typing data}
\usage{
clean_hla_class1(data, var_1, var_2)
}
\arguments{
\item{data}{Data frame containing HLA typing information.}

\item{var_1}{HLA on allele 1.}

\item{var_2}{HLA on allele 2.}
}
\value{
Cleaned data frame with standardized HLA Class I data in original columns
}
\description{
This function processes raw HLA Class I typing data,
removing inconsistent formatting and unnecessary symbols to ensure a standardized allele format.
It also imputes homozygosity at loci where one allele is missing.
}
