% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hla_class2_dp.R
\name{clean_hla_class2_dp}
\alias{clean_hla_class2_dp}
\title{Clean and standardize messy HLA Class II - DP typing data}
\usage{
clean_hla_class2_dp(data, var_1, var_2)
}
\arguments{
\item{data}{Data frame containing HLA typing information.}

\item{var_1}{HLA on allele 1.}

\item{var_2}{HLA on allele 2.}
}
\value{
Cleaned data frame with standardized HLA Class II data in original columns
}
\description{
This function processes raw HLA Class II typing data,
removing inconsistent formatting and unnecessary symbols to ensure a standardized allele format.
It also imputes homozygosity at loci where one allele is missing.
}
\examples{
dat <-  read.csv(system.file("extdata/example", "HLA_Clean_test.csv", package = "tidy_hla"))
re <- clean_hla_class1(dat, recip_dpa1_1, recip_dpa1_2)
}
