% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_mism.R
\name{eval_mism}
\alias{eval_mism}
\title{Evaluate HLA Mismatches}
\usage{
eval_mism(data, don_1, don_2, recip_1, recip_2, hmz_cnt = 1)
}
\arguments{
\item{data}{A data frame containing donor and recipient HLA alleles.}

\item{don_1}{Column name for the first donor allele.}

\item{don_2}{Column name for the second donor allele.}

\item{recip_1}{Column name for the first recipient allele.}

\item{recip_2}{Column name for the second recipient allele.}

\item{hmz_cnt}{Default value is 1. Used to set a homozygosity counter (not used in calculations in this version).}
}
\value{
A modified data frame with mismatches calculated between donor and recipient alleles.
}
\description{
Calculates the number of mismatches between donor and recipient HLA alleles.
The function standardizes input values, aligns missing data,
and assesses mismatches for two sets of alleles per donor and recipient.
}
\examples{
dat <-  read.csv(system.file("extdata/example", "HLA_Clean_test.csv", package = "tidy_hla"))
re <- eval_mism(dat, don_a_1, don_a_2, recip_a_1, recip_a_2)
}
