% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_data.R
\docType{data}
\name{hla_data}
\alias{hla_data}
\title{HLA raw data}
\format{
Data frame with 2 columns and 10
\describe{
\item{data}{A dataframe},
\item{var_1}{Character HLA typing value 1},
\item{var_2}{Character HLA typing value 2}
}
}
\usage{
hla_data
}
\description{
Subject identifier and HLA typing data for Recipients and Donors
}
\examples{
# example code
   hla_data

}
\keyword{datasets}
